# pkg/api/static/api-spec.yaml
openapi: 3.0.0
info:
  title: DNSPulse ConfigHUB API
  version: 1.0.0
  description: DNSPulse Configuration Management REST API
servers:
  - url: http://example.com/api
    description: Production server
paths:
...
  /api/config/general/main:
    get:
      tags:
        - Main Configuration
      summary: Get Main Configuration
      operationId: getMainConfig
      responses:
        '200':
          description: Successful retrieval of the main configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralConfig'
    patch:
      tags:
        - Main Configuration
      summary: Update Main Configuration
      operationId: updateMainConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneralConfig'
      responses:
        '200':
          description: Main configuration updated successfully
        '400':
          description: Invalid request format
    put:
      tags:
        - Main Configuration
      summary: Replace Main Configuration
      operationId: replaceMainConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneralConfig'
      responses:
        '200':
          description: Main configuration replaced successfully
        '400':
          description: Invalid request format
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GeneralConfig:
      type: object
      properties:
        path:
          type: string
          description: Path to the configuration file.
      example:
        path: "/config/general.yaml"
...
security:
  - BearerAuth: []